requirements.txt
django
djangorestframework
streamlit
langgraph
langchain
langchain-openai
psycopg2-binary
redis
requests
fastapi
uvicorn